# Use v2 sets when recovering bindings if available