@michaelklishin 2022-06-17
We will drop Erlang 23 support soon enough, so the check wouldn't be necessary.