@michaelklishin 2022-06-17
@gomoripeti sorry, I did not mean we can do this right away, only that this solution won't last for too long. We have to support Erlang 23 for another month or so.