@lhoguin 2022-06-17
The sets code in 24.0 is better but it was improved again in subsequent 24.x releases, so the benchmark results might differ depending on OTP version.