@gomoripeti 2022-06-17
true that with these remaining 2 lines of change this PR does not make much sense. Its better to consider all sets in RabbitMQ code base together and consider/measure where it makes sense to use v2 and when.
(This was just a side product of my tinkering which I thought I'd ask for some feedback)