@michaelklishin 2022-06-17
Once we drop Erlang 23 support, we can revisit what set types we use.